Eliminate all remaining gcc-15 warnings
- Suppress -Wdiscarded-qualifiers in test files that assign MINC
  string literal macros to char* arrays (inherent API limitation)
- Fix sign-conversion in ezminc fixed_vec loops
- Fix size_t/int conversion in ezminc simple_volume stride computation
- Suppress -Wdiscarded-qualifiers for VIO_STR initializers in
  output_mnc2.c and volumes.c
- Fix ncdimdef sign-conversion in minc_1_rw.cpp

GCC-15 with -Wall -Wextra -Wpedantic -Wformat=2 -Wconversion
-Wsign-conversion -Wdouble-promotion -Wundef -Wshadow -Wcast-qual
-Wcast-align -Wwrite-strings -Wredundant-decls -Wmissing-include-dirs
-Wmissing-prototypes -Wstrict-prototypes -Wold-style-definition
-Wnested-externs now produces 0 warnings.

+++ ezminc/minc_io_fixed_vector.h
@@ -207,14 +207,14 @@ namespace minc
 
     fixed_vec( const fixed_vec& b )
     {
-      for(unsigned int i=0;i<static_cast<unsigned int>(dim);i++)
+      for(int i=0;i<dim;i++)
         c[i]=b[i];
     }
 
     //! assignment operator, copies contents
     fixed_vec<dim,I>& operator=(const fixed_vec<dim,I>& b)
     {
-      for(unsigned int i=0;i<static_cast<unsigned int>(dim);i++)
+      for(int i=0;i<dim;i++)
         c[i]=b[i];
       return *this;
     }
